+++ pando/__main__.py
@@ -2,7 +2,7 @@
 python -m pando
 ===============
 
-Aspen ships with a server (wsgiref.simple_server) that is
+Pando ships with a server (wsgiref.simple_server) that is
 suitable for development and testing.  It can be invoked via:
 
     python -m pando
@@ -10,7 +10,7 @@ suitable for development and testing.  It can be invoked via:
 though even for development you'll likely want to specify a
 project root, so a more likely incantation is:
 
-    ASPEN_PROJECT_ROOT=/path/to/wherever python -m pando
+    PANDO_PROJECT_ROOT=/path/to/wherever python -m pando
 
 For production deployment, you should probably deploy using a higher
 performance WSGI server like Gunicorn, uwsgi, Spawning, or the like.
@@ -56,8 +56,8 @@ logging_cfg = {
 if __name__ == '__main__':
     logging.config.dictConfig(logging_cfg)
     port = int(os.environ.get('PORT', '8080')) # get the port, defaulting to 8080
-    host = os.environ.get('ASPEN_HOST', '0.0.0.0') # get the IP to bind to, or default to all
-    project_root = os.environ.get('ASPEN_PROJECT_ROOT', None)
+    host = os.environ.get('PANDO_HOST', '0.0.0.0') # get the IP to bind to, or default to all
+    project_root = os.environ.get('PANDO_PROJECT_ROOT', None)
     log_dammit("Greetings, program! Now serving on http://{0}:{1}/.".format(host, port))
     website = website.Website(project_root=project_root)
     make_server(host, port, website).serve_forever()